Dodano wiecej logow

+++ BESTTECH/src/main/java/com/aplikacjazespolowa/BESTTECH/controllers/ProductsController.java
@@ -1,25 +1,30 @@
 package com.aplikacjazespolowa.BESTTECH.controllers;
 
+import com.aplikacjazespolowa.BESTTECH.models.*;
 import jakarta.annotation.PostConstruct;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import com.aplikacjazespolowa.BESTTECH.models.Produkt;
-import com.aplikacjazespolowa.BESTTECH.models.ProduktRepository;
-import com.aplikacjazespolowa.BESTTECH.models.Kategoria;
-import com.aplikacjazespolowa.BESTTECH.models.KategoriaRepository;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 
 @Controller
 @RequestMapping("/products")
 public class ProductsController {
     private final ProduktRepository produktRepository;
     private final KategoriaRepository kategoriaRepository;
+
+    @Autowired
+    LogsRepository logsRepository;
+    @Autowired
+    private HttpServletRequest request;
 
     public ProductsController(ProduktRepository produktRepository, KategoriaRepository kategoriaRepository) {
         this.produktRepository = produktRepository;
@@ -91,9 +96,16 @@ public class ProductsController {
                                @RequestParam Integer stanMagazynowy,
                                @RequestParam String marka,
                                @RequestParam Integer kategoriaID,
-                               Model model) {
-        Kategoria kategoria = kategoriaRepository.findById(kategoriaID).orElse(null);
-        if (kategoria != null) {
+                               Model model,RedirectAttributes redirectAttributes, HttpServletRequest request) {
+
+        // pobranie kategorii
+        Optional<Kategoria>kOpt = kategoriaRepository.findById(kategoriaID);
+        if(kOpt.isEmpty()){
+            redirectAttributes.addFlashAttribute("error","Kategoria o ID " + kategoriaID + " nie zostala znaleziona");
+            return "redirect://products/addproduct/";
+        }
+        Kategoria kategoria=kOpt.get();
+
             Produkt produkt = new Produkt();
             produkt.setNazwa(nazwa);
             produkt.setOpis(opis);
@@ -103,17 +115,35 @@ public class ProductsController {
             produkt.setDataDodania(new Date()); // Ustawiamy bieżącą datę
             produkt.setKategoria(kategoria);
 
-            produktRepository.save(produkt); // Zapisujemy produkt do bazy
-        }
+            //log
+            String currentUser=request.getUserPrincipal().getName();
+            logsRepository.save(new LogsSystem("Dodano nowy produkt: " + nazwa ,currentUser,"INFO"));
 
+            produktRepository.save(produkt); // Zapisujemy produkt do bazy
+
+        redirectAttributes.addFlashAttribute("message","Produkt zostal dodany.");
         return "redirect:/products/showproducts"; // Przekierowanie do strony z produktami
     }
 
 
     //problem z usuwaniem kategorii, gdy istnieje produkt z daną kategorią !
     @PostMapping("/deleteproduct")
-    public String usunProdukt(@RequestParam Integer id) {
+    public String usunProdukt(@RequestParam Integer id,HttpServletRequest request, RedirectAttributes redirectAttributes) {
+        //pobranie produktu
+        Optional<Produkt> opt=produktRepository.findById(id);
+        if(opt.isEmpty()){
+            redirectAttributes.addFlashAttribute("error","Produkt o ID " + id + " nie zostal znaleziony");
+            return "redirect://products/showproduct/";
+        }
+        Produkt produkt = opt.get();
+
+        //zapis log o edycji produktu
+        String currentUser=request.getUserPrincipal().getName();
+        Integer deletedProductID=produkt.getProduktID();
+        logsRepository.save(new LogsSystem("Usunieto produkt" + deletedProductID, currentUser,"WARN"));
+
         produktRepository.deleteById(id);
+        redirectAttributes.addFlashAttribute("message","Produkt został usuniety");
         return "redirect:/products/showproducts";
     }
 
@@ -134,19 +164,30 @@ public class ProductsController {
     public String edytujProdukt(@RequestParam Integer id, @RequestParam String nazwa, @RequestParam String opis,
                                 @RequestParam Float cena, @RequestParam String marka,
                                 @RequestParam Integer kategoriaID, @RequestParam Integer stanMagazynowy,
-                                Model model) {
+                                Model model, HttpServletRequest request, RedirectAttributes redirectAttributes) {
 
-        Produkt produkt = produktRepository.findById(id).orElse(null);
-        if (produkt == null) {
-            model.addAttribute("error", "Produkt o ID " + id + " nie został znaleziony.");
-            return "products/showproducts";
+        //znajdywanie produktu
+        Optional<Produkt> opt = produktRepository.findById(id);
+        if(opt.isEmpty()){
+            redirectAttributes.addFlashAttribute("error","Produkt o ID " + id + "nie został znaleziony");
+            return "redirect:/products/showproducts";
         }
+        Produkt produkt = opt.get();
 
-        Kategoria kategoria = kategoriaRepository.findById(kategoriaID).orElse(null);
-        if (kategoria == null) {
-            model.addAttribute("error", "Kategoria o ID " + kategoriaID + " nie została znaleziona.");
-            return "products/editProduct";
+        // pobranie kategorii
+        Optional<Kategoria>kOpt = kategoriaRepository.findById(kategoriaID);
+        if(kOpt.isEmpty()){
+            redirectAttributes.addFlashAttribute("error","Kategoria o ID " + kategoriaID + " nie zostala znaleziona");
+            return "redirect://products/editproduct/" + id;
         }
+        Kategoria kategoria=kOpt.get();
+
+
+        //zapis log o edycji produktu
+        String currentUser=request.getUserPrincipal().getName();
+        Integer editProductID=produkt.getProduktID();
+        logsRepository.save(new LogsSystem("Zmieniono produkt" + editProductID, currentUser,"INFO"));
+
         produkt.setNazwa(nazwa);
         produkt.setOpis(opis);
         produkt.setCena(cena);
@@ -155,6 +196,7 @@ public class ProductsController {
         produkt.setStanMagazynowy(stanMagazynowy);
         produktRepository.save(produkt);
 
+        redirectAttributes.addFlashAttribute("message","Produkt został zaktualizowany");
         return "redirect:/products/showproducts";
     }
 
@@ -174,8 +216,22 @@ public class ProductsController {
 
         //problem z usuwaniem kategorii, gdy istnieje produkt z daną kategorią !
         @PostMapping("/deletecategory")
-        public String usunKategorie (@RequestParam Integer id){
+        public String usunKategorie (@RequestParam Integer id, HttpServletRequest request, RedirectAttributes redirectAttributes){
+            Optional<Kategoria> kOpt=kategoriaRepository.findById(id);
+            if(kOpt.isEmpty()){
+                redirectAttributes.addFlashAttribute("error",
+                        "Kategoria o ID " + id + " nie została znaleziona");
+                return "redirect:/products/showcategories";
+            }
+            Kategoria kategoria=kOpt.get();
+
+            //log
+            String currentUser=request.getUserPrincipal().getName();
+            String deletedCategoryName=kategoria.getNazwa();
+            logsRepository.save(new LogsSystem("Usunieto kategorie " + deletedCategoryName, currentUser,"WARN"));
+
             kategoriaRepository.deleteById(id);
+            redirectAttributes.addFlashAttribute("message","Kategoria \"" + deletedCategoryName + "\" została usunieta");
             return "redirect:/products/showcategories";
         }
 
@@ -190,13 +246,25 @@ public class ProductsController {
             return "products/editCategory";
         }
         @PostMapping("/editcategory")
-        public String edytujKategorie (@RequestParam Integer id, @RequestParam String nazwa, @RequestParam String opis){
-            Kategoria kategoria = kategoriaRepository.findById(id).orElse(null);
-            if (kategoria != null) {
+        public String edytujKategorie (@RequestParam Integer id, @RequestParam String nazwa, @RequestParam String opis,HttpServletRequest request, RedirectAttributes redirectAttributes){
+
+            Optional<Kategoria> kopt=kategoriaRepository.findById(id);
+            if(kopt.isEmpty()){
+                redirectAttributes.addFlashAttribute("error","Kategorii o id: " + id + "nie mozna edytowac");
+                return "redirect:/products/showcategories";
+            }
+            Kategoria kategoria = kopt.get();
+
+            //log
+            String currentUser=request.getUserPrincipal().getName();
+            String editCategoryName=kategoria.getNazwa();
+            logsRepository.save(new LogsSystem("Zaktualizowano kategorię" + editCategoryName, currentUser,"INFO"));
+
                 kategoria.setNazwa(nazwa);
                 kategoria.setOpis(opis);
                 kategoriaRepository.save(kategoria);
-            }
+
+            redirectAttributes.addFlashAttribute("message","Kategoria zostala zaktualizowana");
             return "redirect:/products/showcategories";
         }
 
@@ -206,13 +274,25 @@ public class ProductsController {
         }
 
         @PostMapping("/addcategory")
-        public String dodajKategorie (@RequestParam String nazwa, @RequestParam String opis, Model model){
+        public String dodajKategorie (@RequestParam String nazwa, @RequestParam String opis, Model model,HttpServletRequest request, RedirectAttributes redirectAttributes){
+
+        //tworzenie nowej kategorii
             Kategoria kategoria = new Kategoria();
             kategoria.setNazwa(nazwa);
             kategoria.setOpis(opis);
-            kategoriaRepository.save(kategoria);
-            model.addAttribute("message", "Kategoria dodana pomyślnie!");
-            return "products/addCategory";
+
+
+            //zapisanie log przed zapisem do bazy, zapisanie po raz 1
+            Kategoria saved = kategoriaRepository.save(kategoria);
+
+            //log
+            String CurrentUser=request.getUserPrincipal().getName();
+            String addingCategory=saved.getNazwa();
+            logsRepository.save(new LogsSystem("Dodano nowa kategorie:  " + addingCategory, CurrentUser,"INFO"));
+
+
+            redirectAttributes.addFlashAttribute("message","Dodano nowa kategorie: " + addingCategory);
+            return "/products/showcategories";
         }
